attempt to map physical phone orientation from expo sensor, to camera orientation in 3d r3f scene
why is this so hard lol. will have to do a lot of thinking on how to best use the pitch and roll values from accelerometer, along with the magnetometer for direction. and gyroscope too for relative velocity to smooth things up. including stuff like how to get N to flip to S when the phone is tilted backwards. currently using a colourful 3d sphere to represent where the canvas camera is pointing, but it is not accurate and super scuffed at the moment. will take another stab at this later, need to rest.